dev_django/requirements.txt
Django>=4.2.21,<5.0
djangorestframework>=3.14.0,<4.0
pillow>=9.5.0,<10.0
gunicorn>=21.2.0,<22.0
whitenoise>=6.5.0,<7.0
dj-database-url>=2.1.0,<3.0
psycopg2-binary>=2.9.9,<3.0
